--- src/components/AlertConfirm/index.ts
@@ -1,1 +1,0 @@
-export { Alert, AlertProps, Confirm, ConfirmProps } from './AlertConfirm';
