+++ src/components/Dialogs/Alert/index.ts
@@ -1,0 +1,1 @@
+export { Alert, AlertProps } from './Alert';
