+++ src/components/Dialogs/Confirm/__tests__/Confirm.spec.js
@@ -1,23 +1,29 @@
 import React from 'react';
 import toJson from 'enzyme-to-json';
-import { Alert } from '../AlertConfirm';
+import { Confirm } from '../Confirm';
 
-describe('Alert', () => {
+describe('Confirm', () => {
     const mockOnAccept = jest.fn();
+    const mockOnCancel = jest.fn();
     let wrapper;
 
     beforeEach(() => {
         wrapper = mount(
-            <Alert isOpen acceptButton={{ onClick: mockOnAccept, label: 'OK' }} ariaHideApp={false}>
-                Body of the alert
-            </Alert>
+            <Confirm
+                isOpen
+                acceptButton={{ onClick: mockOnAccept, label: 'OK' }}
+                cancelButton={{ onClick: mockOnCancel, label: 'Cancel' }}
+                ariaHideApp={false}
+            >
+                Body of the confirm
+            </Confirm>
         );
     });
 
     it('should render correctly', () => {
         expect(toJson(wrapper)).toMatchSnapshot();
         expect(wrapper.find('Heading')).toHaveLength(0);
-        expect(wrapper.find('Button')).toHaveLength(1);
+        expect(wrapper.find('Button')).toHaveLength(2);
     });
     it('should render correctly with title', () => {
         wrapper.setProps({ title: 'Title' });
@@ -27,7 +33,11 @@ describe('Alert', () => {
         expect(wrapper.find('Heading')).toHaveLength(1);
     });
     it('should call onAccept cb when button is clicked', () => {
-        wrapper.find('Button').simulate('click');
+        wrapper.find('Button').at(1).simulate('click');
         expect(mockOnAccept).toHaveBeenCalledTimes(1);
+    });
+    it('should call onAccept cb when button is clicked', () => {
+        wrapper.find('Button').at(0).simulate('click');
+        expect(mockOnCancel).toHaveBeenCalledTimes(1);
     });
 });
