+++ src/components/Dialogs/Confirm/index.ts
@@ -1,0 +1,1 @@
+export { Confirm, ConfirmProps } from './Confirm';
