+++ src/components/Dialogs/Dialog/Dialog.scss
@@ -1,6 +1,6 @@
-@import '../../themes/oneui/oneui.scss';
+@import '../../../themes/oneui/oneui.scss';
 
-.Alert {
+.Dialog {
     width: 400px;
 
     &__content {
