+++ src/components/Dialogs/Dialog/index.ts
@@ -1,0 +1,1 @@
+export { Dialog, DialogProps, ButtonProps } from './Dialog';
