+++ src/components/Dialogs/index.ts
@@ -1,0 +1,2 @@
+export { Alert, AlertProps } from './Alert';
+export { Confirm, ConfirmProps } from './Confirm';
